rag-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
chromadb>=0.4.18
openai>=1.6.0
anthropic>=0.8.0
ruamel.yaml>=0.18.5
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
sentence-transformers>=2.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
